MCPJVM-283: Post-epic cleanup is incomplete: legacy MCP Tool entrypoints remain registered after consolidation work
