+++ docs/how-it-works/README.md
@@ -64,7 +64,7 @@ Return the exact request plan and probe verification steps.
 
 1. The orchestrator resolves the project root and passes `projectRootAbs`
 2. `artifact_management` with `artifactType=project_context` and `action=validate` optionally validates the scoped project context for that root
-3. `probe_recipe_create` is called with `projectRootAbs`, `intentMode=line_probe`, exact FQCN in `classHint`, method/line hints, optional `apiBasePath`, and auth context
+3. `route_synthesis` with `action=create_recipe` is called with `projectRootAbs`, `intentMode=line_probe`, exact FQCN in `classHint`, method/line hints, optional `apiBasePath`, and auth context
 4. The tool performs code-based route inference through synthesizer plugins backed by the generic JVM AST request-mapping resolver, returning `executionReadiness`, `requestCandidates`, `inferredTarget`, and `selectedMode`
 5. Framework adapters resolve annotation semantics (for example Spring `@RequestParam`, `@PathVariable`, `@RequestBody`) into normalized parameter metadata before HTTP materialization
 6. Generic HTTP transport materializes path/query/body templates from normalized parameters and applies optional project fixture profile overrides from `.mcp-java-dev-tools/request-template.properties`
@@ -108,8 +108,8 @@ Return endpoint-level HTTP results and any probe-verifiable evidence.
 
 1. The orchestrator resolves the API project and passes `projectRootAbs`
 2. `artifact_management` with `artifactType=project_context` and `action=validate` optionally validates the scoped project context
-3. For each route under the controller, `probe_recipe_create` is called with the exact FQCN (and shared `apiBasePath` when applicable) to produce executable request candidates and auth/readiness diagnostics
-4. If `probe_recipe_create` returns `resultType=report`, that route is treated as fail-closed; use the compact execution metadata and diagnostics for routing decisions
+3. For each route under the controller, `route_synthesis` with `action=create_recipe` is called with the exact FQCN (and shared `apiBasePath` when applicable) to produce executable request candidates and auth/readiness diagnostics
+4. If `route_synthesis` with `action=create_recipe` returns `resultType=report`, that route is treated as fail-closed; use the compact execution metadata and diagnostics for routing decisions
 5. The orchestrator fires regression HTTP requests route-by-route with bearer auth and records outcomes
 6. Probe verification is only applied when strict line targets are available for an endpoint
 7. For probe-eligible endpoints: `probe` with `action=reset` -> execute HTTP request -> `probe` with `action=wait_for_hit` (or `action=status`) confirms runtime line execution
@@ -147,7 +147,7 @@ For every failed or flagged run, generate a reproducible recipe and include runt
 
 1. Run the full Use Case 2 regression flow and collect all endpoint outcomes
 2. Filter endpoints into `failed` or `flagged` sets (non-2xx, contract mismatch, probe-miss, etc.)
-3. For each failed/flagged endpoint, call `probe_recipe_create` to produce a focused rerun recipe tied to the observed failure context
+3. For each failed/flagged endpoint, call `route_synthesis` with `action=create_recipe` to produce a focused rerun recipe tied to the observed failure context
 4. If strict line verification is possible, run `probe` with `action=reset` -> targeted HTTP rerun -> `probe` with `action=wait_for_hit` / `action=status`
 5. If a capture preview exists, call `probe` with `action=capture` and attach the evidence to that endpoint's recipe package
 6. If the runtime route can't be uniquely validated during rerun, emit a fail-closed pushback artifact, not speculative guidance
