+++ test/integrations/spring/social_platform/create_recipe/mapper_plugin_unavailable.it.ts
@@ -34,12 +34,15 @@ test("create_recipe REPORT: preserves mapper_plugin_unavailable from AST resolve
   if (!runtime || !mcp) throw new Error("runtime/mcp not initialized");
 
   const recipe = (await mcp.client.callTool({
-    name: "probe_recipe_create",
+    name: "route_synthesis",
     arguments: {
-      projectRootAbs: path.join(socialPlatformRootAbs, "post-service", "post-app"),
-      classHint: postControllerFqcn,
-      methodHint: "listPosts",
-      intentMode: "regression",
+      action: "create_recipe",
+      input: {
+        projectRootAbs: path.join(socialPlatformRootAbs, "post-service", "post-app"),
+        classHint: postControllerFqcn,
+        methodHint: "listPosts",
+        intentMode: "regression",
+      },
     },
   })) as any;
