+++ test/tools/spring/recipe-synthesis-contract.test.ts
@@ -1,7 +1,7 @@
 const assert = require("node:assert/strict");
 const test = require("node:test");
 
-const { generateRecipe } = require("@/tools/core/recipe_generate/domain");
+const { generateRecipe } = require("@/tools/core/route_synthesis/shared/recipe_generation.util");
 
 const okAuth = {
   required: "unknown",
