+++ tools/transport/tools-mcp-server/src/models/inputs/route_synthesis/index.ts
@@ -1,0 +1,10 @@
+export {
+  RouteSynthesisInputSchema,
+  RouteSynthesisRequestSchema,
+  type RouteSynthesisRequest,
+} from "@/models/inputs/route_synthesis/route_synthesis.input.model";
+export {
+  RouteSynthesisActionSchema,
+  ROUTE_SYNTHESIS_ACTION_ALLOWLIST,
+  type RouteSynthesisAction,
+} from "@/models/inputs/route_synthesis/shared/actions.model";
