+++ tools/transport/tools-mcp-server/src/tools/core/route_synthesis/shared/target_inference.util.ts
